git-svn-id: https://opengl-scut.googlecode.com/svn/trunk@459 81c3a017-244d-672b-1dc6-c8d6962b81c2

+++ Code/Scar/Scar/NetworkBase.h
@@ -68,7 +68,7 @@ namespace Network
 		virtual void OtherMessageHandler( unsigned long ip, const PACKAGE& p )	{}
 
 		// Í¬²½tcp·¢ËÍ
-		virtual void TcpSendTo( int ip, int port, const PACKAGE& p )
+		virtual void TcpSendTo( unsigned long ip, int port, const PACKAGE& p )
 		{
 			using namespace boost::asio;
